backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Backend for CoonSpect"

authors = []

requires-python = ">=3.13"
dependencies = [
    "aiohttp>=3.13.3",
    "alembic>=1.18.4",
    "asyncpg>=0.31.0",
    "celery>=5.6.2",
    "fastapi>=0.133.0",
    "passlib[argon2]>=1.7.4",
    "psycopg>=3.3.3",
    "pydantic-settings>=2.13.1",
    "pyjwt>=2.11.0",
    "python-multipart>=0.0.22",
    "redis>=7.2.0",
    "requests>=2.32.5",
    "sqlalchemy[asyncio]>=2.0.47",
    "uvicorn[standard]>=0.41.0",
]

[build-system]
requires = ["uv_build>=0.9.28,<0.10.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "ipdb>=0.13.13",
    "ipython>=9.10.0",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
]
lint = [
    "mypy>=1.19.1",
    "ruff>=0.15.2",
]

# Tools

[tool.uv]
package = false

[tool.ruff]
line-length = 88
target-version = "py313"
exclude = ["alembic"]
src = ["src"]

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "ASYNC"]
fixable = ["ALL"]
ignore = []

[tool.ruff.lint.per-file-ignores]
"src/app/db/models/*" = ["F821"]

[tool.mypy]
python_version = "3.13"
strict = true
ignore_missing_imports = true
plugins = ["sqlalchemy.ext.mypy.plugin"]

[[tool.mypy.overrides]]
module = ["passlib.*"]
ignore_missing_imports = true
